# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/enterprise/reports/dashboard": {
      "get": {
        "operationId": "EnterpriseReportingController_dashboard",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
    "/api/v1/enterprise/reports/impact": {
      "get": {
        "operationId": "EnterpriseReportingController_enterpriseImpact",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
    "/api/v1/enterprise/reports/groups/{id}/impact": {
      "get": {
        "operationId": "EnterpriseReportingController_groupImpact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
    "/api/v1/enterprise/reports/clusters/{id}/impact": {
      "get": {
        "operationId": "EnterpriseReportingController_clusterImpact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
    "/api/v1/enterprise/reports/territories/{id}/impact": {
      "get": {
        "operationId": "EnterpriseReportingController_territoryImpact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
    "/api/v1/enterprise/reports/sites/{id}/impact": {
      "get": {
        "operationId": "EnterpriseReportingController_siteImpact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
    "/api/v1/enterprise/reports/breakdown": {
      "get": {
        "operationId": "EnterpriseReportingController_enterpriseBreakdown",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One level down from the Enterprise — impact per group.",
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
    "/api/v1/enterprise/reports/groups/{id}/breakdown": {
      "get": {
        "operationId": "EnterpriseReportingController_groupBreakdown",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
    "/api/v1/enterprise/reports/clusters/{id}/breakdown": {
      "get": {
        "operationId": "EnterpriseReportingController_clusterBreakdown",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
    "/api/v1/enterprise/reports/sites/rankings": {
      "get": {
        "operationId": "EnterpriseReportingController_rankings",
        "parameters": [
          {
            "name": "scopeType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scopeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Site league table plus active / inactive / never-used / deactivated counts.",
        "tags": [
          "EnterpriseReporting"
        ]
      }
    },
...
  },
  "info": {
    "title": "Saveful for Business API",
    "description": "Food rescue platform API. All routes are prefixed `/api/v1`.\n\n**Authentication.** Every endpoint requires a bearer token unless marked otherwise. Obtain one from `POST /auth/login`, then send it as `Authorization: Bearer <token>`.\n\n**Enterprise endpoints** additionally require the organisation to be on the Enterprise plan, and are gated on the caller’s Enterprise role and scope. A 403 carrying `MISSING_PERMISSION` means the role is wrong; one carrying `OUTSIDE_SCOPE` means the role is right but the target sits outside what the caller can reach.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}